Abstracted repeated logic and implemented it throghout classes
